RavenDB-1290 Size of memory needed to allocate buffers for free pages should take into account that internally UnmanagedBits stores the bits as integers

+++ Voron/Impl/IBufferProvider.cs
@@ -5,7 +5,7 @@ namespace Voron.Impl
 	public interface IBufferProvider : IDisposable
 	{
 		void IncreaseSize(long size);
-		unsafe byte* GetBufferPointer();
+		unsafe int* GetBufferPointer();
 		void Release();
 	}
 }
